sudah membuat logic crud tabel feedback

+++ database/migrations/2021_12_09_051627_create_feedback_table.php
@@ -20,7 +20,7 @@ class CreateFeedbackTable extends Migration
             $table->uuid('film_id');
             $table->foreign('film_id')->references('id')->on('films');
             $table->text('ulasan');
-            $table->integer('rating');
+            $table->float('rating');
             $table->timestamps();
         });
     }
